Move Preprocessed Columns Struct to Preprocessed File
