+++ stwo_cairo_prover/crates/prover/src/components/memory/memory_address_to_id/component.rs
@@ -9,7 +9,7 @@ use stwo_prover::core::fields::qm31::SecureField;
 use stwo_prover::core::fields::secure_column::SECURE_EXTENSION_DEGREE;
 use stwo_prover::core::pcs::TreeVec;
 
-use crate::components::memory::Seq;
+use crate::cairo_air::preprocessed::Seq;
 use crate::relations;
 
 /// Split the (ID , Multiplicity) columns to shorter chunks. This is done to improve the performance
